requirements.txt
# Web framework
fastapi==0.115.5
uvicorn[standard]==0.32.1

# Database
sqlalchemy==2.0.36
psycopg[binary]>=3.1
alembic==1.14.0

# CV parsing
pdfplumber==0.11.4
python-docx==1.1.2

# Config
python-dotenv==1.0.1
python-multipart==0.0.18

# OpenAI (optional – set OPENAI_API_KEY to enable)
openai>=2.0
